requirements.txt
torch
torchvision
numpy==1.26.4
netCDF4
transformers
torchaudio
pretrainedmodels
pyhocon
imageio
opencv-python
imageio-ffmpeg
tensorboard
scipy
scikit-image
ipdb
matplotlib
tqdm
timm>=1.0.9
lpips
chardet
omegaconf
hydra-core
pillow
google-cloud-storage
einops
av
dask
boto3
xarray
h5py
rsatoolbox
moviepy==1.0.3
datalad
pymoten
POT
albumentations
h5netcdf
gdown
scikit-learn
progressbar2
pandas
psutil
pyyaml
huggingface-hub
wandb
brainscore-core
brainscore-brainio

# Audio/video processing
decord
librosa
soundfile

# Deep learning utilities
pytorch-lightning
geotorch
torchdiffeq

# mmengine/mmaction for SlowFast models
mmengine
mmaction2==1.1.0
mmcv>=2.0.0,<2.1.0
importlib-metadata<5

# git model imports
r3m @ git+https://github.com/facebookresearch/r3m.git#egg=r3m
selavi @ git+https://github.com/YingtianDt/selavi.git
robustbench @ git+https://github.com/RobustBench/robustbench.git
phys_readouts@ git+https://github.com/thekej/phys_readouts.git#egg=phys_readouts
avid_cma @ git+https://github.com/YingtianDt/AVID-CMA.git
gdt_model @ git+https://github.com/YingtianDt/GDT.git
openstl @ git+https://github.com/YingtianDt/OpenSTL.git
S3D_HowTo100M @ git+https://github.com/YingtianDt/S3D_HowTo100M
vonenet @ git+https://github.com/dicarlolab/vonenet.git
jepa @ git+https://github.com/thekej/jepa.git
